requirements.txt
django
django-rest-framework
statsd
psycopg2-binary
watchdog
lxml
kafka-python